refactor!: move coldObservableProvider to util-rxjs package
