+++ packages/util-rxjs/src/index.ts
@@ -5,4 +5,5 @@ export * from './passthrough';
 export * from './finalizeWithLatest';
 export * from './concatAndCombineLatest';
 export * from './shareRetryBackoff';
+export * from './coldObservableProvider';
 export * from './types';
