+++ packages/wallet/src/services/DelegationTracker/RewardAccounts.ts
@@ -28,7 +28,7 @@ import {
 import { RetryBackoffConfig } from 'backoff-rxjs';
 import { TrackedStakePoolProvider } from '../ProviderTracker';
 import { TxWithEpoch } from './types';
-import { coldObservableProvider } from '../';
+import { coldObservableProvider } from '@cardano-sdk/util-rxjs';
 import findLast from 'lodash/findLast';
 import isEqual from 'lodash/isEqual';
 import uniq from 'lodash/uniq';
