+++ packages/wallet/src/services/util/index.ts
@@ -1,5 +1,4 @@
 export * from './persistentTrackerSubjects';
 export * from './equals';
 export * from './trigger';
-export * from './coldObservableProvider';
 export * from './connectionStatusTracker';
